Renamed core project from OwlVoiceAssistant to VoiceAssistant
