--- app/src/javampd.properties
@@ -1,3 +1,0 @@
-monitor.delay=60
-monitor.playlist.multiplier=0
-monitor.track.multiplier=0
