--- app/src/main/java/OwlVoiceAssistant/CalBacks/IntentCallBack.java
@@ -1,7 +1,0 @@
-package OwlVoiceAssistant.CalBacks;
-
-import OwlVoiceAssistant.TextToIntent.Intent;
-
-public interface IntentCallBack {
-    void intentParsed(Intent intent);
-}
