+++ lib/src/main/java/VoiceAssistant/CalBacks/IntentCallBack.java
@@ -1,0 +1,7 @@
+package VoiceAssistant.CalBacks;
+
+import VoiceAssistant.TextToIntent.Intent;
+
+public interface IntentCallBack {
+    void intentParsed(Intent intent);
+}
