+++ lib/src/main/java/VoiceAssistant/TextToIntent/Command.java
@@ -1,4 +1,4 @@
-package OwlVoiceAssistant.TextToIntent;
+package VoiceAssistant.TextToIntent;
 
 
 import java.util.*;
