+++ lib/src/main/java/VoiceAssistant/TextToIntent/TTI.java
@@ -1,4 +1,4 @@
-package OwlVoiceAssistant.TextToIntent;
+package VoiceAssistant.TextToIntent;
 
 import com.google.gson.Gson;
 import org.apache.logging.log4j.LogManager;
