+++ lib/src/main/java/example/configuration.properties
