+++ lib/src/test/java/VoiceAssistant/AppTest.java
@@ -1,10 +1,9 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package OwlVoiceAssistant;
+package VoiceAssistant;
 
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
 
 class AppTest {
     @Test void appHasAGreeting() {
