+++ settings.gradle
@@ -7,5 +7,5 @@
  * in the user manual at https://docs.gradle.org/7.2/userguide/multi_project_builds.html
  */
 
-rootProject.name = 'OwlVoiceAssistant'
-include('app')
+rootProject.name = 'voice-assistant'
+include('lib')
